requirements.txt
cffi==2.0.0
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
fonttools==4.61.1
greenlet==3.3.2
imbalanced-learn==0.14.1
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
nvidia-nccl-cu12==2.29.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pycparser==3.0
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
SQLAlchemy==2.0.46
threadpoolctl==3.6.0
typing_extensions==4.15.0
xgboost==3.2.0
